fix: now cipher function gets an array of types instead of one type
