+++ package.json
@@ -1,0 +1,20 @@
+{
+  "name": "ciphering_cli_tool",
+  "version": "1.0.0",
+  "description": "NodeJS 2021 Q4 task",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/fpastl/Ciphering_CLI_Tool.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/fpastl/Ciphering_CLI_Tool/issues"
+  },
+  "homepage": "https://github.com/fpastl/Ciphering_CLI_Tool#readme"
+}
